Update SlicerOpenLIFU hash in CMakeLists.txt (#14)
Updates CMakeLists.txt to reference a merged commit that has a fix to an
issue where a residual part of the placeholder guided mode widget
remained after being swapped out. That is, a purple box was clouding the
view even after the purple placeholder widget was replaced. This issue
is described in
[SlicerOpenLIFU#236](https://github.com/OpenwaterHealth/SlicerOpenLIFU/issues/236).
While the referenced commit does temporarily fix the issue,
SlicerOpenLIFU#236 was not closed at the time of this commit.